(refactor) individualize components
+ FetchUser function globally exported
+ Searchscreen is now a component
+ useFetchRepos is now a global hook
